pyproject.toml
[project]
name = "project1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.3.28",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.11",
    "python-dotenv>=1.2.1",
]
